+++ backends/connect/srcs/schemas/ws.ts
@@ -1,0 +1,4 @@
+import { Type } from "@sinclair/typebox";
+
+// Placeholder for future query params
+export const readyQuerySchema = Type.Object({});
