--- backends/connect/srcs/types/consts.ts
@@ -1,20 +1,0 @@
-import { WebSocket } from "@fastify/websocket";
-
-export const RoomStatus = {
-  WAITING: "00",
-  OCCUPIED: "01",
-} as const;
-
-export const ActionTypes = {
-  CREATE: "00",
-  JOIN: "01",
-  HEART_BEAT: "02",
-  KEY_SIGNAL: "03",
-} as const;
-
-export const EventTypes = {
-  CREATED_ROOM: "00",
-  JOINED_ROOM: "01",
-  HEALTH_CHECK: "02",
-  GAME_FIELD: "03",
-} as const;
